добил основные запросы, теперь проект билдится

+++ Application/UseCase/Tenants/CreateTenant/CreateTenantUseCase.cs
@@ -6,10 +6,10 @@ namespace Application.UseCase.Tenants.CreateTenant;
 
 public class CreateTenantUseCase(ICreateTenantStorage storage)
 {
-    public async Task<Result<int>> CreateTenant(CreateTenantRequest request)
+    public async Task<Result<long>> CreateTenant(CreateTenantRequest request)
     {
         var result = await storage.CreateTenant(request);
 
-        return Result<int>.Success(result);
+        return Result<long>.Success(result);
     }
 }
