+++ Infrastructure/Database/Initializers/Seeders/Tenants.cs
@@ -4,7 +4,7 @@ namespace Infrastructure.Database.Initializers.Seeders;
 
 public partial class DataSeeder
 {
-    public async Task<Result<int>> CreateTenantByName(string tenantName)
+    public async Task<Result<long>> CreateTenantByName(string tenantName)
     {
         return await createTenant.CreateTenant(new(Name: tenantName));
     }
